Swap reserved ranges for Generic Desktop and Digitizer pages
Commit 8a16361 appears to have inadvertantly stored the reserved ranges of
the Digitzer page in the Generic Desktop page, and vice-versa. This looks
like a simple case of trimming the wrong bits from the wrong file after
copying the 'id_set_range_desktop.m4' as a prototype. This commit swaps
the two ranges back so they are associated with the correct page.