Don't break when defining usages greater than INT32_MAX with GCC
While the C standard defines enum constants to have type "int", it
leaves the type of the enum itself implementation defined. GCC defines
that the type will be unsigned as long as there are no negative values
and signed otherwise[1]. The "hidrd_usage_valid" function implicitly
relies on this fact since it compares usages to the HIDRD_USAGE_MAX
which is not a constant of the enum, but rather a macro equal to
UINT32_MAX. If the "hidrd_usage" enum contains a constant greater
than INT32_MAX, GCC changes the enum's type to signed which results
in the value HIDRD_USAGE_MAX being interpreted as -1 instead.

This patch has "hidrd_usage_valid" cast the min, max, and value to be
tested to uint32_t before performing the comparison to ensure that it
is always performed as intended. This seems to be sufficeient to fix
"make check", but related issues may still be lurking.

[1]: https://gcc.gnu.org/onlinedocs/gcc/Structures-unions-enumerations-and-bit-fields-implementation.html