Fix printing of multiple type descriptions
The loop currently used to build a type description string does not work
properly when multiple types are possible. hidrd_usage_type_idx_name is
only ever called once, causing the first description to be repeated.
Calling the function in the 'else' case ensures that we see all of the
type descriptions.